separate build and install steps
Build the package using makepkg as the non-root user, but then install
the package as root using pacman. This avoids any sudo password prompts
from makepkg attempting to become root to install the package.

Passing the use_unsafe_shell argument to run_command seems to be
necessary to get it to properly glob the wildcards in the compiled
package name, which is necessary since we don't know what the full
filename will be.